仮で作成していたモデルを統合しました
